chore: replace previous eslint impl completely
