+++ testlib/src/test/java/com/diffplug/spotless/npm/EslintFormatterStepTest.java
@@ -27,7 +27,6 @@ import com.diffplug.common.collect.ImmutableMap;
 import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.ResourceHarness;
 import com.diffplug.spotless.StepHarnessWithFile;
-import com.diffplug.spotless.TestProvisioner;
 import com.diffplug.spotless.tag.NpmTest;
 
 @NpmTest
@@ -60,7 +59,6 @@ class EslintFormatterStepTest {
 
 			final FormatterStep formatterStep = EslintFormatterStep.create(
 					devDependenciesForRuleset.get(ruleSetName),
-					TestProvisioner.mavenCentral(),
 					projectDir(),
 					buildDir(),
 					null,
@@ -103,7 +101,6 @@ class EslintFormatterStepTest {
 
 			final FormatterStep formatterStep = EslintFormatterStep.create(
 					devDependenciesForRuleset.get(ruleSetName),
-					TestProvisioner.mavenCentral(),
 					projectDir(),
 					buildDir(),
 					null,
@@ -160,7 +157,6 @@ class EslintFormatterStepTest {
 
 			final FormatterStep formatterStep = EslintFormatterStep.create(
 					EslintStyleGuide.TS_XO_TYPESCRIPT.mergedWith(EslintFormatterStep.defaultDevDependenciesForTypescript()),
-					TestProvisioner.mavenCentral(),
 					projectDir(),
 					buildDir(),
 					null,
